backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
django-environ
django-extensions
drf-yasg
django-leaflet
Pillow
# GeoDjango dependencies
# Note: On Windows, GDAL must be installed separately
